test : (refactor) 전체 Service 기능 테스트(Stub) 리팩토링
- 요청과 동일할 수 있도록 Id와 request가 분리되어 있도록 변경
- doAnswer를 이용하여 save 처리를 새롭게 생성할 수 있도록 변경

+++ src/main/java/kitchenpos/domain/Product.java
@@ -14,10 +14,10 @@ public class Product {
         this.price = price;
     }
 
-    public Product(Long id, String name, int price) {
+    public Product(Long id, String name, BigDecimal price) {
         this.id = id;
         this.name = name;
-        this.price = BigDecimal.valueOf(price);
+        this.price = price;
     }
 
     public Long getId() {
